// src/app/shared/models/enums/security-role.enum.ts
export enum SecurityRole {
  ROLE_SIGNATORY = 'ROLE_SIGNATORY', // 0
  ROLE_AS_CREATOR = 'ROLE_AS_CREATOR', // 1 - Co-Requestor
  ROLE_AS_SIGNATORY = 'ROLE_AS_SIGNATORY', // 2 - Observer
  ROLE_TEMPLATE_OWNER = 'ROLE_TEMPLATE_OWNER', // 3
  ROLE_TEMPLATE_USER = 'ROLE_TEMPLATE_USER', // 4
  ROLE_AS_CONSUMER = 'ROLE_AS_CONSUMER', // 5
  ROLE_CREATOR = 'ROLE_CREATOR' // 6
}